🚑 fix: 가맹점 정보 업데이트 시 접근 권한 체크하도록 수정

+++ src/main/java/com/x1/frans/user/query/dto/HqUserDepartmentDTO.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @Setter
 public class HqUserDepartmentDTO {
 
-    private Long departmentId;
+    private Integer departmentId;
     private String departmentName;
     private Integer depth;
 }
